+++ rows/plugins/dicts.py
@@ -17,23 +17,42 @@
 
 from __future__ import unicode_literals
 
+from itertools import chain
+
 from rows.plugins.utils import create_table
 
 
-def import_from_dicts(data, *args, **kwargs):
-    """Import data from a list of dicts."""
-    headers = set()
-    for row in data:
-        headers.update(row.keys())
-    headers = sorted(list(headers))
+def import_from_dicts(data, samples=None, *args, **kwargs):
+    """Import data from a iterable of dicts
 
-    data = [[row.get(header, None) for header in headers] for row in data]
+    The algorithm will use the `samples` first `dict`s to determine the field
+    names (if `samples` is `None` all `dict`s will be used).
+    """
 
-    meta = {'imported_from': 'dicts', }
-    return create_table([headers] + data, meta=meta, *args, **kwargs)
+    data = iter(data)
+
+    cached_rows, headers = [], []
+    for index, row in enumerate(data, start=1):
+        cached_rows.append(row)
+
+        for key in row.keys():
+            if key not in headers:
+                headers.append(key)
+
+        if samples and index == samples:
+            break
+
+    data_rows = (
+        [row.get(header, None) for header in headers]
+        for row in chain(cached_rows, data)
+    )
+
+    kwargs["samples"] = samples
+    meta = {"imported_from": "dicts"}
+    return create_table(chain([headers], data_rows), meta=meta, *args, **kwargs)
 
 
 def export_to_dicts(table, *args, **kwargs):
-    """Export rows.Table to dict."""
-    return [{key: getattr(row, key) for key in table.field_names}
-            for row in table]
+    """Export a `rows.Table` to a list of dicts"""
+    field_names = table.field_names
+    return [{key: getattr(row, key) for key in field_names} for row in table]